utils/datasets: Move get_dataset_files and get_dataset_module to utils/datasets
